Improved update operations
Evidently, the package is available via the operation instance composed
in the event, and both the operation and the event are immutable. As
such, this requires a different approach to updating:

- We need to grab the operation from the provided PackageEvent
- We need to retrieve the initial and target packages from the operation
- We need to build a new package event for the uninstall operation that
  includes a new UninstallOperation instance composing the initial
  package.
- We need to build a new package event for the install operation that
  includes a new InstallOperation instance composing the target
  package.

This patch adds a method for creating a new PackageEvent based on the
original, but containing the new operation provided, and updates the
`onPostPackageUpdate()` method to call on this method, providing an
appropriate operation composing the appropriate package instance.

The results have been integration tested in an existing application, and
I have verified that the assets were updated.