Fixes composer installation errors due to problematic configuration files
If a module configuration file contains elements that depend on
autoloading and/or elements that are clearly not configuration, calling
`include` on it can lead to failure to complete a composer
installation/removal action.

This patch modifies the logic to tokenize the configuration file and
look for problematic elements; when encountered, the (un)installer will skip
including that configuration file, and instead emit a warning.