Do not emit warnings for class pseudo-constant or files without asset_manager configuration
If a configuration file does not contain `asset_manager` configuration,
we do not need to do anything with it in the first place, and do not
need to emit warnings.

If the file contains any class pseudo-constant usage (e.g.
`AssetInstallerFactory::class`), we do not need to flag the file as
unusable, as PHP will resolve these regardless of autoloading.

As such, we'll only have issues with class constants, static members,
and cases where we are calling `new` or extending or implementing
classes or interfaces.